Move colors and effects into separate files.
